+++ src/app/login/styles.tsx
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  /* Estilos adicionais se necessário */
-`;